% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_importation.R
\name{add_chat_transcript}
\alias{add_chat_transcript}
\title{Add chat messages to a transcript}
\usage{
add_chat_transcript(
  transcript_data,
  chat_transcript,
  start_time,
  chat_format = "webex"
)
}
\arguments{
\item{transcript_data}{A data frame containing the transcript.}

\item{chat_transcript}{A string with the path to the chat file or a data
frame containing the chat data.}

\item{start_time}{The start time of the meeting. A string in the format
"HH:MM AM/PM" or "HH:MM:SS".}

\item{chat_format}{A string indicating the online meeting platform used to
generate the chat file. Only "webex" is supported at the moment.}
}
\value{
A data frame containing the transcript with the chat messages added.
}
\description{
This function adds chat messages written during an online/hybrid meeting to
the transcript.
}
