% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{build_ids_from_agenda}
\alias{build_ids_from_agenda}
\title{Builds IDs from each agenda item session and titles}
\usage{
build_ids_from_agenda(agenda)
}
\arguments{
\item{agenda}{A list containing the agenda items.}
}
\value{
A character vector containing the IDs for each agenda item.
}
\description{
Builds IDs from each agenda item session and titles
}
