% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_agenda_summary_tree_consistency}
\alias{check_agenda_summary_tree_consistency}
\title{Check consistency of agenda and summary tree ids}
\usage{
check_agenda_summary_tree_consistency(agenda, summary_tree)
}
\arguments{
\item{agenda}{A list containing the agenda or a file path to it.}

\item{summary_tree}{A list containing the summary tree or a file path to it.}
}
\value{
Nothing, will throw an error if the agenda and summary tree are not
consistent.
}
\description{
Check consistency of agenda and summary tree ids
}
