% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_install_dependencies}
\alias{check_and_install_dependencies}
\title{Check and install dependencies}
\usage{
check_and_install_dependencies(deps)
}
\arguments{
\item{deps}{A character vector of package names.}
}
\value{
Nothing.
}
\description{
Checks if a list of packages are installed and installs them if not. Trigger
an error if the user chooses not to install a package.
}
