% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_summary_tree_consistency}
\alias{check_summary_tree_consistency}
\title{Validate summary tree id consistency}
\usage{
check_summary_tree_consistency(summary_tree)
}
\arguments{
\item{summary_tree}{A list containing the summary tree or a file path to it.}
}
\value{
Nothing, will throw an error if the summary tree is not consistent.
}
\description{
Validate summary tree id consistency
}
