% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{convert_agenda_times}
\alias{convert_agenda_times}
\title{Convert agenda clock time to seconds from start}
\usage{
convert_agenda_times(
  agenda,
  event_start_time = getOption("minutemaker_event_start_time")
)
}
\arguments{
\item{agenda}{The agenda of the meeting, that is, a list of agenda elements
each with a session name, a title, speaker and moderator lists, type of
talk and start and end times. Alternatively, an agenda element directly.}

\item{event_start_time}{The start time of the event in the HH:MM(:SS)( AM/PM)
format. Will be used to compute the time lapse in seconds from the start of
the event. If \code{NULL}, defaults to the start time of the first element in
the agenda, with a warning. Cannot be \code{NULL} if agenda is actually an
agenda element.}
}
\value{
The agenda (or agenda element) with all the times in seconds from the
start.
}
\description{
Users can provide events' times in the agenda following the HH:MM(:SS)(
AM/PM) format, but then most function require it as seconds from the start to
match events to the transcript
}
