% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_generation.R
\name{.gen_cntx_blocks_by_attr}
\alias{.gen_cntx_blocks_by_attr}
\title{Index extracted blocks by a chosen attribute.}
\usage{
.gen_cntx_blocks_by_attr(blocks, attr)
}
\arguments{
\item{blocks}{List of blocks returned by \code{.gen_cntx_extract_xml_blocks()}.}

\item{attr}{Attribute name to index on.}
}
\value{
Named list mapping attribute values to block content.
}
\description{
Index extracted blocks by a chosen attribute.
}
\keyword{internal}
