% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_generation.R
\name{.gen_cntx_is_plain_text}
\alias{.gen_cntx_is_plain_text}
\title{Detect whether a file looks like plain text.}
\usage{
.gen_cntx_is_plain_text(path, sample_bytes = 8192, control_threshold = 0.3)
}
\arguments{
\item{path}{Path to a file to sample.}

\item{sample_bytes}{Number of bytes to sample for detection.}

\item{control_threshold}{Maximum proportion of control bytes tolerated.}
}
\value{
Logical value indicating whether the file is likely plain text.
}
\description{
Detect whether a file looks like plain text.
}
\keyword{internal}
