% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_generation.R
\name{.gen_cntx_read_materials}
\alias{.gen_cntx_read_materials}
\title{Read documentation materials from a folder.}
\usage{
.gen_cntx_read_materials(material_dir)
}
\arguments{
\item{material_dir}{Directory holding documentation files.}
}
\value{
Named list of file contents (character scalars).
}
\description{
Read documentation materials from a folder.
}
