% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_interpolation_env}
\alias{.make_interpolation_env}
\title{Build interpolation environment for path_exists messages}
\usage{
.make_interpolation_env(path, caller_env, type)
}
\arguments{
\item{path}{A character string with the path for interpolation.}

\item{caller_env}{The calling environment for interpolation.}

\item{type}{A character string indicating the path type.}
}
\value{
An environment with \verb{_path}, \verb{_basename}, and \code{type} values.
}
\description{
Build interpolation environment for path_exists messages
}
\keyword{internal}
