% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{entity_extractor}
\alias{entity_extractor}
\title{Extract entities from a text}
\usage{
entity_extractor(
  text,
  entities = c("people", "acronyms", "organizations", "concepts"),
  prompt_only = FALSE,
  ...
)
}
\arguments{
\item{text}{The text from which to extract the entities.}

\item{entities}{A character vector with the entities to extract. Can be
"people", "acronyms", "organizations", and "concepts". Default is all of
them.}

\item{prompt_only}{If TRUE, only the prompt is returned, the LLM is not
interrogated. Default is FALSE.}

\item{...}{Additional arguments passed to the \code{llmR::prompt_llm}
function.}
}
\value{
A vector with the entities found in the text.
}
\description{
This function takes a text and extracts entities from it. The entities can be
people, acronyms, organizations, and concepts. The function returns a vector
with the entities found in the text. Can be useful to build vocabularies for
the LLMs starting from an event description or a transcript.
}
