% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{format_agenda}
\alias{format_agenda}
\title{Formats an agenda into a human-readable text}
\usage{
format_agenda(
  agenda,
  event_start_time = getOption("minutemaker_event_start_time")
)
}
\arguments{
\item{agenda}{A list containing the agenda items. It is contains a number of
information about each talk, such as the session, title, speakers,
moderators, start and ending times.}

\item{event_start_time}{If agenda timings are in seconds, the starting time
is needed to convert them to actual clock time. If \code{NULL} it will use the
timing as reported in the agenda.}
}
\value{
A string containing the formatted agenda, invisibly.
}
\description{
The agenda functions returns a machine-readable agenda. This function takes
an agenda in R list format as generated by the \code{infer_agenda_from_transcript}
function and formats it into a human-readable text.
}
