% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{generate_agenda_element_prompt}
\alias{generate_agenda_element_prompt}
\title{Generate the prompt to extract an agenda element details from a transcript}
\usage{
generate_agenda_element_prompt(transcript_segment, args)
}
\arguments{
\item{transcript_segment}{A segment of the transcript to be used for
extracting the details of an agenda element. Can be a character vector
representing the data in CSV format or a data frame.}

\item{args}{A list of arguments to be passed to the prompt template. They can
include: event_description and vocabulary.}
}
\value{
A prompt used by \code{infer_agenda_from_transcript()}.

,
}
\description{
This function is used by \code{infer_agenda_from_transcript()} to generate a
prompt for extracting the details of an agenda element from a transcript.
}
