% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{generate_recording_details}
\alias{generate_recording_details}
\title{Generate Recording Details}
\usage{
generate_recording_details(
  agenda_element = NULL,
  type = NULL,
  session = NULL,
  title = NULL,
  speakers = NULL,
  moderators = NULL
)
}
\arguments{
\item{agenda_element}{A list containing the details of the recording. If this
is provided, the other parameters are ignored.}

\item{type}{The type of the talk/meeting, e.g., "conference talk",
"workshop", "panel discussion", "conference welcome message", etc.}

\item{session}{The session in which the talk/meeting took place.}

\item{title}{The title of the talk/meeting.}

\item{speakers}{The main speakers of the talk/meeting as a vector.}

\item{moderators}{The moderators of the talk/meeting as a vector.}
}
\value{
A formatted string containing the details of the recording, or NULL
if all parameters are NULL.
}
\description{
This function generates a formatted string containing details about a
recording. The details include the type of talk/meeting, the session in which
it took place, the title of the talk/meeting, the main speakers, and the
moderators. All parameters are optional.
}
