% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{generate_rolling_aggregation_prompt}
\alias{generate_rolling_aggregation_prompt}
\title{Generate one rolling window summarisation prompt}
\usage{
generate_rolling_aggregation_prompt(summaries, args)
}
\arguments{
\item{summaries}{A character vector of summaries of the transcript segments
generated through a "rolling window" approach}

\item{args}{A list of arguments to be passed to the prompt template.}
}
\value{
A summarisation prompt used by \code{summarise_transcript()}.
}
\description{
This function is used by \code{summarise_transcript()} to aggregate the summaries
generated by the "rolling window" summarisation of a single transcript.
}
