% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{generate_summarisation_prompt}
\alias{generate_summarisation_prompt}
\title{Generate a transcript summarisation prompt}
\usage{
generate_summarisation_prompt(transcript, args)
}
\arguments{
\item{transcript}{A transcript text, for example generated by
\code{extract_text_from_transcript()}.}

\item{args}{A list of arguments to be passed to the prompt template.}
}
\value{
A summarisation prompt used by \code{summarise_transcript()}.
}
\description{
This function is used by \code{summarise_transcript()} to generate a summarisation
prompt for a single transcript.
}
