% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_processing.R, R/speech_to_text.R
\name{is_audio_file_sane}
\alias{is_audio_file_sane}
\title{Check if an audio file is corrupted using ffmpeg}
\usage{
is_audio_file_sane(audio_file)

is_audio_file_sane(audio_file)
}
\arguments{
\item{audio_file}{The path to the audio file to check.}
}
\value{
TRUE if the file is valid, FALSE if it is corrupted.

TRUE if the file is valid, FALSE if it is corrupted.
}
\description{
This function uses ffmpeg to check if an audio file is corrupted. It does
this by trying to decode the audio and sending it to a null sink. If ffmpeg
returns an error, the file is considered corrupted.

This function uses ffmpeg to check if an audio file is corrupted. It does
this by trying to decode the audio and sending it to a null sink. If ffmpeg
returns an error, the file is considered corrupted.
}
