% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_silent}
\alias{is_silent}
\title{Check if transcript segments are silent}
\usage{
is_silent(segments)
}
\arguments{
\item{segments}{A vector of transcript segments.}
}
\value{
A logical vector indicating if the segment is silent.
}
\description{
Check if transcript segments are silent, e.g., \link{...}, NA, or empty strings.
}
