% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{merge_transcripts}
\alias{merge_transcripts}
\title{Merge two transcripts}
\usage{
merge_transcripts(transcript_x, transcript_y, import_diarization = TRUE)
}
\arguments{
\item{transcript_x}{A data frame containing the transcript to which the
second transcript will be merged.}

\item{transcript_y}{A data frame containing the transcript to be merged in.}

\item{import_diarization}{A boolean indicating whether the diarization
information should be kept from the second transcript.}
}
\value{
A data frame containing the merged transcript.
}
\description{
Merges two transcripts into one, overwriting segments that are empty in the
first transcript with the corresponding segments from the second transcript.
Also imports the diarization information if present.
}
