% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutemaker.R
\docType{package}
\name{minutemaker}
\alias{minutemaker}
\alias{minutemaker-package}
\title{minutemaker}
\description{
Package which uses LLM models to generate minutes from audio recordings or transcriptions of meetings.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
