% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_event_time}
\alias{parse_event_time}
\title{Parse time if it is a string}
\usage{
parse_event_time(time, format = c("R", "T"))
}
\arguments{
\item{time}{A time string in the "HH:MM(:SS)( AM/PM)", format.}

\item{format}{A character vector of formats to try.}
}
\value{
A POSIXct value representing the time or the original value if it is
numeric.
}
\description{
Parse time if it is a string
}
