% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_to_text.R
\name{run_in_terminal}
\alias{run_in_terminal}
\title{Run a command in a terminal window}
\usage{
run_in_terminal(cmd, args)
}
\arguments{
\item{cmd}{The command to run.}

\item{args}{The arguments to pass to the command as a character vector.}
}
\value{
The terminal/command line window ID.
}
\description{
This function is useful since it gives more control than just running
commands from R directly. For example it allows to check the output of the
command in real time.
}
\examples{

# Run a command in a terminal window
run_in_terminal("ls", args = c("-l", "-a"))
}
