% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{set_prompts}
\alias{set_prompts}
\title{Set the technical prompts that power the LLM model}
\usage{
set_prompts(
  base_task = NULL,
  transcript_template = NULL,
  event_description_template = NULL,
  recording_details_template = NULL,
  vocabulary_template = NULL,
  diarization_template = NULL,
  diarization_instructions = NULL,
  audience_template = NULL,
  audience = NULL,
  summarisation_template = NULL,
  summarisation_sections = NULL,
  output_template = NULL,
  output_instructions = NULL,
  force = TRUE
)
}
\arguments{
\item{base_task}{The base task of the AI secretary.}

\item{transcript_template}{The template for the transcript.}

\item{event_description_template}{The template for the event description.}

\item{recording_details_template}{The template for the recording details.}

\item{vocabulary_template}{The template for the vocabulary.}

\item{diarization_template}{The template for the diarization instructions.}

\item{diarization_instructions}{Default instructions for the diarization.}

\item{audience_template}{The template for the audience.}

\item{audience}{Description of the audience and the summary focus.}

\item{summarisation_template}{The template for the summarisation
instructions.}

\item{summarisation_sections}{The sections to include in the summary.}

\item{output_template}{The template for the output instructions.}

\item{output_instructions}{Instructions for the output, i.e., how to produce
and format the summary.}

\item{force}{If TRUE, overwrite existing prompts.}
}
\value{
Nothing. The prompts are set as options in the R environment.
}
\description{
This function helps setting the prompts for the LLM that is tasked with
summarising meetings and conference presentations. The prompts are set as
options in the R environment. If \code{NULL} is passed as an argument, the default
prompts are used.`
}
