% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_to_text.R
\name{split_audio}
\alias{split_audio}
\title{Split an audio file into segments of a specified duration}
\usage{
split_audio(
  audio_file,
  segment_duration = 60,
  output_folder = file.path(dirname(audio_file), "recording_parts")
)
}
\arguments{
\item{audio_file}{The path to the audio file to split.}

\item{segment_duration}{The duration of each segment in minutes.}

\item{output_folder}{The folder to save the segments in.}
}
\value{
Nothing, but saves the segments to files.
}
\description{
Split an audio file into segments of a specified duration
}
