% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{summarise_full_meeting}
\alias{summarise_full_meeting}
\title{Summarise a full meeting}
\usage{
summarise_full_meeting(
  transcript_data,
  agenda,
  output_file = "event_summary.R",
  event_description = NULL,
  event_audience = NULL,
  vocabulary = NULL,
  consider_diarization = TRUE,
  summarisation_sections = NULL,
  diarization_instructions = NULL,
  output_instructions = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{transcript_data}{The transcript data of the meeting as a data frame.}

\item{agenda}{The agenda of the meeting, that is, a list of agenda elements
each with a session name, a title, speaker and moderator lists, type of
talk and start and end times. Alternatively, the path to an R file
containing such a list.}

\item{output_file}{The file to save the results to as a list. Default is
"event_summary.R".}

\item{event_description}{The description of the event See
\code{summarise_transcript} for more details.}

\item{event_audience}{The audience of the event See \code{summarise_transcript}
for more details.}

\item{vocabulary}{The vocabulary used in the meeting. See
\code{summarise_transcript} for more details.}

\item{consider_diarization}{Whether to take into account the diarization of
the transcript. Default is TRUE See \code{summarise_transcript} for more
details.}

\item{summarisation_sections, diarization_instructions, output_instructions}{Specific instructions necessary to build the summarisation prompt. See
\code{summarise_transcript} for more details and run \code{get_prompts()} to see the
defaults.}

\item{overwrite}{Whether to overwrite existing summaries. Default is FALSE.}

\item{...}{Additional arguments passed to \code{interrogate_llm} function, such as
the LLM provider.}
}
\value{
The result tree of the meeting summary. Also saves the results in the
output file as side effect.
}
\description{
This function summarises a full event comprising multiple meetings/talks
described in an \code{agenda}. It generates a summary for each talk in the meeting
and saves the results in an output file. It's built on top of
\code{summarise_transcript}.
}
