% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_to_numeric}
\alias{time_to_numeric}
\title{Transform a time string to numeric}
\usage{
time_to_numeric(time)
}
\arguments{
\item{time}{A time string in the "HH:MM(:SS)( AM/PM)", format.}
}
\value{
A numeric value representing the time.
}
\description{
Transform a time string to numeric
}
