% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_operations.R
\name{tokenize_text}
\alias{tokenize_text}
\title{Tokenize Text for Text Analysis}
\usage{
tokenize_text(text)
}
\arguments{
\item{text}{A character vector containing the text to be tokenized.}
}
\value{
A list of tokens extracted from the text.
}
\description{
This function processes and tokenizes a given text. It converts the text to
lowercase, removes special characters and extra spaces, and then splits the
text into tokens.
}
