% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_calls.R
\name{use_azure_llm}
\alias{use_azure_llm}
\title{Use Azure Language Model}
\usage{
use_azure_llm(
  body,
  deployment_id = getOption("minutemaker_azure_deployment_gpt"),
  resource_name = getOption("minutemaker_azure_resource_gpt"),
  api_key = getOption("minutemaker_azure_api_key_gpt"),
  api_version = getOption("minutemaker_azure_api_version")
)
}
\arguments{
\item{body}{The body of the request.}

\item{deployment_id}{Azure deployment ID for the language model. Obtained
from R options.}

\item{resource_name}{Azure resource name. Obtained from R options.}

\item{api_key}{API key for the Azure language model service. Obtained from R
options.}

\item{api_version}{API version for the Azure language model service. Obtained
from R options.}
}
\value{
The function returns the response from the Azure API.
}
\description{
Sends a request to the Azure API for language model completions using the
parameters in the \code{body} argument. This function requires specific Azure
configurations (deployment ID, resource name, API key, and API version) set
in the R options.
}
