% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_to_text.R
\name{use_azure_whisper_model}
\alias{use_azure_whisper_model}
\title{Use Azure Whisper Model for Speech-to-Text}
\usage{
use_azure_whisper_model(
  audio_file,
  language = "en",
  initial_prompt = "",
  temperature = 0,
  resource_name = getOption("minutemaker_azure_resource_name"),
  deployment_id = getOption("minutemaker_azure_deployment_whisper"),
  api_key = getOption("minutemaker_azure_api_key"),
  api_version = getOption("minutemaker_azure_api_version")
)
}
\arguments{
\item{audio_file}{The path to the audio file to transcribe.}

\item{language}{The language of the input audio in mp3/wav format. Should be
less than 25 MB.}

\item{temperature}{number (optional) The sampling temperature between 0 and
\enumerate{
\item Defaults to 0.
}}

\item{resource_name}{The name of the Azure resource to use.}

\item{deployment_id}{The deployment ID of the Azure resource to use.}

\item{api_key}{The API key to use for the Azure resource.}

\item{api_version}{The API version to use for the Azure resource.}

\item{prompt}{Text to guide the model's style or continue a previous segment.}
}
\value{
A list with the full transcript and the transcription by segments.
}
\description{
This function sends an audio file to the Azure Whisper API for transcription.
See https://learn.microsoft.com/en-us/azure/ai-services/openai/reference.
}
