% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_to_text.R
\name{use_openai_whisper_model}
\alias{use_openai_whisper_model}
\title{Use OpenAI Whisper Model for Speech-to-Text}
\usage{
use_openai_whisper_model(
  audio_file,
  language = "en",
  initial_prompt = "",
  temperature = 0,
  api_key = getOption("minutemaker_openai_api_key")
)
}
\arguments{
\item{audio_file}{The path to the audio file to transcribe.}

\item{language}{The language of the input audio in mp3/wav format. Should be
less than 25 MB.}

\item{temperature}{number (optional) The sampling temperature between 0 and
\enumerate{
\item Defaults to 0.
}}

\item{api_key}{The OpenAI API key. If not provided, it will be retrieved
from the global options.}

\item{prompt}{Text to guide the model's style or continue a previous segment.}
}
\value{
A list with the full transcript and the transcription by segments.
}
\description{
This function sends an audio file to the Azure Whisper API for transcription.
See https://platform.openai.com/docs/api-reference/audio/createTranscription
}
