% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_agenda}
\alias{validate_agenda}
\title{Validates an agenda}
\usage{
validate_agenda(agenda, ...)
}
\arguments{
\item{agenda}{A list containing the agenda or a file path to it.}

\item{...}{Additional arguments to be passed to \code{validate_agenda_element}.}
}
\value{
A boolean indicating whether the agenda is valid.
}
\description{
Checks if the agenda is a list (or a file path to a list) and if it is not
empty and if all its elements are valid.
}
\examples{
validate_agenda(list(
 list(
  session = "Session 1",
  title = "Opening Session",
  speakers = "John Doe",
  moderators = "Jane Doe",
  type = "conference talk",
  from = "09:00 AM",
  to = "10:00 AM"
 ),
 list()
 ), session = TRUE, title = TRUE, speakers = TRUE, moderators = TRUE,
 type = TRUE, from = TRUE, to = TRUE)
 #> [1] FALSE # Because the second element is empty

}
