% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{validate_agenda_element}
\alias{validate_agenda_element}
\title{Validates an agenda element}
\usage{
validate_agenda_element(
  agenda_element,
  session = FALSE,
  title = FALSE,
  speakers = FALSE,
  moderators = FALSE,
  type = FALSE,
  from = FALSE,
  to = FALSE
)
}
\arguments{
\item{agenda_element}{A list containing the agenda elements.}

\item{session}{A boolean indicating whether the \code{session} item should be
present.}

\item{title}{A boolean indicating whether the \code{title} item should be present.}

\item{speakers}{A boolean indicating whether the \code{speakers} item should be
present.}

\item{moderators}{A boolean indicating whether the \code{moderators} item should
be present.}

\item{type}{A boolean indicating whether the \code{type} item should be present.}

\item{from}{A boolean indicating whether the \code{from} item should be present.}

\item{to}{A boolean indicating whether the \code{to} item should be present.}
}
\value{
A boolean indicating whether the agenda element is valid.
}
\description{
Validates an agenda element
}
