% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda_management.R
\name{convert_agenda_times}
\alias{convert_agenda_times}
\title{Convert agenda clock time to seconds from start and vice versa}
\usage{
convert_agenda_times(
  agenda,
  convert_to = c("seconds", "clocktime"),
  event_start_time = getOption("minutemaker_event_start_time"),
  conversion_format = "\%T"
)
}
\arguments{
\item{agenda}{The agenda of the meeting, that is, a list of agenda elements
each with a session name, a title, speaker and moderator lists, type of
talk and start and end times. Alternatively, an agenda element directly.}

\item{convert_to}{A string indicating whether to convert the times to seconds
from the start or to clock time. Conversion to seconds will happen only if
the times are in clock time format, and vice versa.}

\item{event_start_time}{The start time of the event in the HH:MM(:SS)( AM/PM)
format or a POSIXct object. Will be used to convert event times to seconds
from the start. If \code{NULL}, defaults to the start time of the first element
in the agenda, with a warning.}

\item{conversion_format}{The format to use when converting to clock time.
E.g., \%T for HH:MM:SS, or \%R for HH:MM, etc... See \code{?strptime} for a list
of supported formats.}
}
\value{
The agenda (or agenda element) with all the times in seconds from the
start.
}
\description{
Users can provide events' times in the agenda following the HH:MM(:SS)(
AM/PM) format or as number of seconds from the start of the event. This
function converts the times to the other format.
}
