% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_manipulation.R
\name{extract_text_from_transcript}
\alias{extract_text_from_transcript}
\title{Extracts a specific talk text from a given transcript data}
\usage{
extract_text_from_transcript(
  transcript_data,
  agenda_element = NULL,
  start_time = NULL,
  end_time = NULL,
  import_diarization = TRUE
)
}
\arguments{
\item{transcript_data}{A data frame containing the transcript data. It must
have the following columns: 'start', 'end' and 'text'}

\item{agenda_element}{A list with 'from' and 'to' fields indicating the start
and end times of the agenda element.}

\item{start_time}{The start time for the extraction.}

\item{end_time}{The end time for the extraction.}

\item{import_diarization}{A boolean indicating whether the speaker should be
mentioned, if present. Will be ignored if the transcript does not contain
speaker information or there is only one speaker.}
}
\value{
A string containing the extracted text.
}
\description{
The specific talk text can be identify using the relative agenda element or
manually inputting the start and end times.
}
