% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{format_summary_tree}
\alias{format_summary_tree}
\title{Formats a summary tree into a human-readable text}
\usage{
format_summary_tree(summary_tree, agenda, output_file = NULL)
}
\arguments{
\item{summary_tree}{A list containing the summary tree, or a string with the
path to a file containing the summary tree.}

\item{agenda}{A list containing the agenda items. It is used to extract a number
of information about each talk, such as the session, title, speakers and
moderators.}

\item{output_file}{A string with the path to the output file. If NULL, the
output is returned invisibly and not written to a file.}
}
\value{
A string containing the formatted summary tree, invisibly. This is
printed to the output file if one is provided.
}
\description{
The summary functions returns a machine-readable summary tree. This function
takes the summary tree and formats it into a human-readable text.
}
