% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_operations.R
\name{generate_glove_model}
\alias{generate_glove_model}
\title{Generate Word Vectors from Text using GloVe Model}
\usage{
generate_glove_model(
  text,
  dimensions = 100,
  iterations = 50,
  overwrite = FALSE
)
}
\arguments{
\item{text}{A character vector of text to be processed.}

\item{dimensions}{The number of dimensions of the word vectors.}

\item{iterations}{The number of iterations to run the GloVe model.}

\item{overwrite}{Whether to overwrite a previously generated model.}
}
\value{
A matrix of word embeddings.
}
\description{
This function takes a text input and uses the \code{text2vec} package's GloVe
model to generate word embeddings.
}
\examples{
\dontrun{
text <- "Sample text for GloVe model generation."
word_vectors <- generate_glove_model(text, dimensions = 5, iterations = 10)
}

}
