% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{get_prompts}
\alias{get_prompts}
\title{Get the current prompts}
\usage{
get_prompts(which = NULL)
}
\arguments{
\item{which}{A vector of one or more prompt names.}
}
\value{
A list containing the current prompts.
}
\description{

This function returns the current prompts as a list. If no argument is
provided, all the prompts are returned. If a vector of prompt names is
provided, only the requested prompts are returned.

}
