% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutemaker-package.R
\docType{package}
\name{minutemaker}
\alias{minutemaker-package}
\alias{minutemaker}
\title{minutemaker}
\description{
Package which uses LLM models to generate minutes from audio recordings or transcriptions of meetings.
}
\author{
\strong{Maintainer}: Angelo D'Ambrosio \email{a.dambrosioMD@gmail.com} (\href{https://orcid.org/0000-0002-2045-5155}{ORCID})

}
