% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{parse_transcript_json}
\alias{parse_transcript_json}
\title{Parse the speech-to-text JSON output into a transcript data frame}
\usage{
parse_transcript_json(
  transcript_json,
  pretty_times = TRUE,
  clock_start_time = getOption("minutemaker_event_start_time")
)
}
\arguments{
\item{transcript_json}{The JSON output of \code{perform_speech_to_text} in
multiple formats: the raw JSON, a list, a path to a JSON file, or a path to
a folder containing JSON files. If a folder is provided, the function will
parse all JSON files in the folder.}

\item{pretty_times}{A boolean indicating whether the start and end times
should be also formatted as hours-seconds-minutes instead of cumulative
seconds.}

\item{clock_start_time}{The start time of the event in the HH:MM(:SS)( AM/PM)
format. Will be used to add the actual clock time to the transcript
segments. If NULL, the clock time will not be added.}
}
\value{
A data frame with the text and start and end time of each segment.
}
\description{
Takes the JSON output of \code{perform_speech_to_text} and parses it into a data
frame with the text and time of each segment.
}
