% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{set_prompts}
\alias{set_prompts}
\title{Set technical prompts that power the LLM model}
\usage{
set_prompts(
  base_task = NULL,
  transcript = NULL,
  vocabulary = NULL,
  diarization_instructions = NULL,
  extra_diarization_instructions = NULL,
  audience_instructions = NULL,
  audience = NULL,
  summarisation_focus = NULL,
  output_format = NULL,
  presentation_details = NULL,
  meeting_description = NULL,
  force = TRUE
)
}
\arguments{
\item{base_task}{The base task of the AI secretary.}

\item{transcript}{The transcript to be summarised.}

\item{vocabulary}{The vocabulary to be used.}

\item{diarization_instructions}{Instructions for diarization.}

\item{extra_diarization_instructions}{Extra instructions for diarization.}

\item{audience_instructions}{Instructions for the audience and the summary
focus.}

\item{audience}{The audience for the summary.}

\item{summarisation_focus}{The focus of the summarisation.}

\item{output_format}{The format of the output.}

\item{presentation_details}{The details of the presentation.}

\item{meeting_description}{The description of the meeting.}

\item{force}{If TRUE, overwrite existing prompts.}
}
\value{
Nothing. The prompts are set as options in the R environment.
}
\description{
This function helps setting the prompts for the LLM that is tasked with
summarising meetings and conference presentations. The prompts are set as
options in the R environment. If \code{NULL} is passed as an argument, the default
prompts are used.`
}
