% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{set_prompts}
\alias{set_prompts}
\title{Set the technical prompts that power the LLM model}
\usage{
set_prompts(
  persona = NULL,
  base_task = NULL,
  aggregate_task_rolling = NULL,
  event_description_template = NULL,
  recording_details_template = NULL,
  transcript_template = NULL,
  aggregate_template_rolling = NULL,
  vocabulary_template = NULL,
  diarization_template = NULL,
  audience_template = NULL,
  summary_template = NULL,
  summary_structure = NULL,
  output_template = NULL,
  output_summarisation = NULL,
  output_rolling_aggregation = NULL,
  force = TRUE
)
}
\arguments{
\item{persona}{The general persona description of the AI secretary.}

\item{base_task}{The base task for a trasncript summarisation.}

\item{aggregate_task_rolling}{The task for the rolling window summary
aggregation}

\item{event_description_template}{The template for inserting the event
description.}

\item{recording_details_template}{The template for inserting the recording
details.}

\item{transcript_template}{The template for inserting the transcript .}

\item{aggregate_template_rolling}{The template for inserting the summaries
for the rolling window aggregation.}

\item{vocabulary_template}{The template for inserting the vocabulary.}

\item{diarization_template}{The template with standard diarization
instructions and the placeholder for extra custom instructions.}

\item{audience_template}{The template for inserting the audience and summary
focus description.}

\item{summary_template}{The template for inserting the summary structure
instructions.}

\item{summary_structure}{The default summary structure instructions.}

\item{output_template}{The general output instruction template, with
placeholders for trascript summarisation/summaries' aggregation output
instruction and extra, user-defined, instructions.}

\item{output_summarisation}{The default output instructions for the
transcript summarisation.}

\item{output_rolling_aggregation}{The default output instructions for the
"rolling window" generated summaries' aggregation.}

\item{force}{If TRUE, overwrite existing prompts.}
}
\value{
Nothing. The prompts are set as options in the R environment.
}
\description{
This function helps setting the prompts for the LLM that is tasked with
summarising meetings and conference presentations. The prompts are set as
options in the R environment. If \code{NULL} is passed as an argument, the default
prompts are used.`
}
