% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_to_text.R
\name{split_audio}
\alias{split_audio}
\title{Split an audio file into segments of a specified duration}
\usage{
split_audio(
  audio_file,
  segment_duration = 40,
  output_folder = file.path(dirname(audio_file), "recording_parts")
)
}
\arguments{
\item{audio_file}{The path to the audio file to split.}

\item{segment_duration}{The duration of each splitted audio file in minutes.
20 minutes equate to more or less 7-8 MB.}

\item{output_folder}{The path to the folder where to save the segments in.}
}
\value{
Nothing, but saves the segments to files.
}
\description{
Some speech-to-text models have a limit on the size of the audio file. For
example, the "Whisper" based models have a limit of 25 MB. This function
splits an audio file into segments of a specified duration and saves them as
mp3 files.
}
