% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{summarise_full_meeting}
\alias{summarise_full_meeting}
\title{Summarise a full meeting}
\usage{
summarise_full_meeting(
  transcript_data,
  agenda,
  method = c("simple", "rolling"),
  output_file = "event_summary.R",
  window_size = 15,
  output_length = 3,
  event_start_time = getOption("minutemaker_event_start_time"),
  event_description = NULL,
  audience = "An audience with understanding of the topic",
  vocabulary = NULL,
  consider_diarization = TRUE,
  summary_structure = get_prompts("summary_structure"),
  extra_diarization_instructions = NULL,
  extra_output_instructions = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{transcript_data}{The transcript data of the meeting as a data frame.}

\item{agenda}{The agenda of the meeting, that is, a list of agenda elements
each with a session name, a title, speaker and moderator lists, type of
talk and start and end times. Alternatively, the path to an R file
containing such a list.}

\item{method}{The method to use to summarise the transcript. Can be either
"simple" or "rolling", with the first being faster and the second
(theoritically) more accurate. See \code{summarise_transcript} for more details.}

\item{output_file}{The file to save the results to as a list. Default is
"event_summary.R".}

\item{window_size}{The size of the window in minutes to use when summarising
the transcript using the "rolling window" method. See
\code{summarise_transcript} for more details.}

\item{output_length}{An indication to the LLM regarding the length of the
output. See \code{summarise_transcript} for more details.}

\item{event_start_time}{The start time of the event in the HH:MM(:SS)( AM/PM)
format. Necessary to convert the agenda times to seconds if the first event
time is not the start time of the event.}

\item{event_description}{The description of the event See
\code{summarise_transcript} for more details.}

\item{audience}{The audience of the event See \code{summarise_transcript}
for more details.}

\item{vocabulary}{The vocabulary used in the meeting. See
\code{summarise_transcript} for more details.}

\item{consider_diarization}{Whether to take into account the diarization of
the transcript. Default is TRUE See \code{summarise_transcript} for more
details.}

\item{summary_structure, extra_diarization_instructions, extra_output_instructions}{Specific instructions necessary to build the summarisation prompt. See
\code{summarise_transcript} for more details and run \code{get_prompts()} to see the
defaults. See \code{summarise_transcript} for more details.}

\item{overwrite}{Whether to overwrite existing summaries. Default is FALSE.}

\item{...}{Additional arguments passed to \code{interrogate_llm} function, such as
the LLM provider.}
}
\value{
The result tree of the meeting summary. Also saves the results in the
output file as side effect.
}
\description{
This function summarises a full event comprising multiple meetings/talks
described in an \code{agenda}. It generates a summary for each talk in the meeting
and saves the results in an output file. It's built on top of
\code{summarise_transcript}.
}
