% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{summarise_full_meeting}
\alias{summarise_full_meeting}
\title{Summarise a full meeting}
\usage{
summarise_full_meeting(
  transcript_data,
  agenda,
  output_file = "event_summary.R",
  meeting_description = NULL,
  meeting_audience = NULL,
  vocabulary = NULL,
  consider_diarization = TRUE,
  extra_diarization_instructions = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{transcript_data}{The transcript data of the meeting as a data frame.}

\item{agenda}{The agenda of the meeting, that is, a list of agenda elements
each with a session name, a title, speaker and moderator lists, type of
talk and start and end times. Alternatively, the path to an R file
containing such a list.}

\item{output_file}{The file to save the results to as a list. Default is
"event_summary.R".}

\item{meeting_description}{The description of the meeting. See
\code{summarise_transcript} for more details.}

\item{meeting_audience}{The audience of the meeting. See
\code{summarise_transcript} for more details.}

\item{vocabulary}{The vocabulary used in the meeting. See
\code{summarise_transcript} for more details.}

\item{consider_diarization}{Whether to take into account the diarization of
the transcript. Default is TRUE See \code{summarise_transcript} for more
details.}

\item{extra_diarization_instructions}{Extra instructions for diarization. See
\code{summarise_transcript} for more details.}

\item{overwrite}{Whether to overwrite existing summaries. Default is FALSE.}

\item{...}{Additional arguments passed to \code{interrogate_llm} function, such as
the LLM provider.}
}
\value{
The result tree of the meeting summary. Also saves the results in the
output file as side effect.
}
\description{
This function summarises a full meeting by processing the transcript data and
the agenda. It generates a summary for each talk in the meeting and saves the
results in an output file. It's built on top of \code{summarise_transcript}.
}
