% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{summarise_transcript}
\alias{summarise_transcript}
\title{Summarise a recording transcript}
\usage{
summarise_transcript(
  transcript,
  event_description = NULL,
  event_details = NULL,
  audience = NULL,
  vocabulary = NULL,
  is_diarized = FALSE,
  extra_diarization_instructions = NULL,
  ...
)
}
\arguments{
\item{transcript}{A character vector representing the transcript to be
summarised.}

\item{event_description}{A description of the event.}

\item{event_details}{Additional details about the event. We suggest using
\code{generate_recording_details} to generate this string.}

\item{audience}{A description the intended audience and the focus of the
summarisation.}

\item{vocabulary}{A character vector of specific vocabulary words, names,
definitions, to help the LLM recognise misspellings and abbreviations.}

\item{is_diarized}{A logical indicating whether the transcript is diarized.
Default is FALSE.}

\item{extra_diarization_instructions}{Additional diarization instructions,
e.g., "Do not report the name of the speaker", "Use only the second name of
the speaker", etc...}

\item{...}{Additional arguments passed to the \code{interrogate_llm} function,
such as the LLM provider.}
}
\value{
A summary of the transcript.
}
\description{
This function takes a transcript and various optional parameters, and returns
a summarised version of the transcript. All parameters apart from the
transcript are optional, and if not provided they will be ignored in the
final prompt.
}
