% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_to_numeric}
\alias{time_to_numeric}
\title{Transform a time string to numeric}
\usage{
time_to_numeric(time, origin = NULL)
}
\arguments{
\item{time}{A time string in the "HH:MM(:SS)( AM/PM)" format, a POSIXct
object, or a number (i.e., of seconds).}

\item{origin}{Same format as \code{time}; if provided, used to compute the
difference in seconds from \code{time}.}
}
\value{
A numeric value representing the difference in seconds from \code{origin}.
}
\description{
Transform a time string to numeric
}
